feat: New infix expression evaluator
